requirements.txt
streamlit==1.39.0
yfinance==0.2.54
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
arabic-reshaper==3.0.0
python-bidi==0.6.0
joblib==1.4.2
keras==3.7.0
tensorflow==2.18.0
scikit-learn

